use float32 instead of double to save memory

+++ hdbscan/dist_metrics.pyx
@@ -14,7 +14,7 @@ np.import_array()  # required in order to use C-API
 
 from libc.math cimport fabs, sqrt, exp, cos, pow, log, acos, M_PI
 
-DTYPE = np.double
+DTYPE = np.float32
 ITYPE = np.intp
